updated model

+++ util.py
@@ -1,11 +1,12 @@
 import os
+import re
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
 import pickle
 import sqlalchemy
 import numpy as np
 import pandas as pd
-
+from sklearn.utils import shuffle
 from wordcloud import WordCloud
 from nltk.corpus import stopwords
 from sqlalchemy import create_engine
@@ -14,6 +15,8 @@ from matplotlib import pyplot as plt
 from nltk.stem import WordNetLemmatizer
 from nltk.tokenize import RegexpTokenizer
 from sklearn.preprocessing import LabelEncoder
+
+from tensorflow.keras.preprocessing.sequence import pad_sequences
 
 from variables import*
 
@@ -32,34 +35,35 @@ def remove_stop_words(stopwords_list,sentence):
     '''
     return [k for k in sentence if k not in stopwords_list]
 
-def preprocess_one(concern):
+def preprocess_one(description):
     '''
         Text preprocess on term text using above functions
     '''
     lemmatizer = WordNetLemmatizer()
     tokenizer = RegexpTokenizer(r'\w+')
     stopwords_list = stopwords.words('english')
-    concern = concern.lower()
-    remove_punc = tokenizer.tokenize(concern) # Remove puntuations
+    description = description.lower()
+    remove_punc = tokenizer.tokenize(description) # Remove puntuations
+    remove_num = [re.sub('[0-9]', '', i) for i in remove_punc] # Remove Numbers
     remove_num = [i for i in remove_punc if len(i)>0] # Remove empty strings
     lemmatized = lemmatization(lemmatizer,remove_num) # Word Lemmatization
     remove_stop = remove_stop_words(stopwords_list,lemmatized) # remove stop words
-    updated_concern = ' '.join(remove_stop)
-    return updated_concern
+    updated_description = ' '.join(remove_stop)
+    return updated_description
 
-def preprocessed_data(concerns):
+def preprocessed_data(descriptions):
     '''
         Preprocess entire terms
     '''
-    updated_concerns = []
-    if isinstance(concerns, np.ndarray) or isinstance(concerns, list):
-        for concern in concerns:
-            updated_concern = preprocess_one(concern)
-            updated_concerns.append(updated_concern)
-    elif isinstance(concerns, np.str_)  or isinstance(concerns, str):
-        updated_concerns = [preprocess_one(concerns)]
+    updated_descriptions = []
+    if isinstance(descriptions, np.ndarray) or isinstance(descriptions, list):
+        for description in descriptions:
+            updated_description = preprocess_one(description)
+            updated_descriptions.append(updated_description)
+    elif isinstance(descriptions, np.str_)  or isinstance(descriptions, str):
+        updated_descriptions = [preprocess_one(descriptions)]
 
-    return np.array(updated_concerns)
+    return np.array(updated_descriptions)
 
 def label_encoding(Y):
     Y = Y.str.strip()
@@ -76,22 +80,23 @@ def label_encoding(Y):
             encoder = pickle.load(handle)
     return encoder.transform(Y)
 
-def create_wordcloud(processed_concerns):
-    long_string = ','.join(list(processed_concerns))
-    wordcloud = WordCloud(
-                        width=1600, 
-                        height=800, 
-                        max_words=200, 
-                        background_color='white',
-                        max_font_size=200, 
-                        random_state=seed
-                        )
-    wordcloud.generate(long_string)
-    plt.imshow(wordcloud, interpolation='bilinear')
-    plt.axis("off")
-    plt.title("WordCloud Distribution of Student Concerns")
-    plt.savefig(wordcloud_path)
-    plt.show()
+def create_wordcloud(processed_descriptions):
+    if not os.path.exists(wordcloud_path):
+        long_string = ','.join(list(processed_descriptions))
+        wordcloud = WordCloud(
+                            width=1600, 
+                            height=800, 
+                            max_words=200, 
+                            background_color='white',
+                            max_font_size=200, 
+                            random_state=seed
+                            )
+        wordcloud.generate(long_string)
+        plt.imshow(wordcloud, interpolation='bilinear')
+        plt.axis("off")
+        plt.title("WordCloud Distribution of Student descriptions")
+        plt.savefig(wordcloud_path)
+        plt.show()
 
 def create_database(engine):
     engine = create_engine(db_url)
@@ -109,8 +114,8 @@ def get_Data():
     create_database(engine)
 
     df = pd.read_sql_table(table_name, engine)
+    df_response = df[['Category', 'Description', 'Book_title' ,'Author' ,'ISBN-10' ,'ISBN-13', 'Cover_link']] 
     df= df[['Description', 'Category']]
-
     description = df['Description'].values
     Category = df['Category']
 
@@ -118,4 +123,38 @@ def get_Data():
     create_wordcloud(X)
 
     Y = label_encoding(Category)
-    return X, Y
+
+    X, Y = shuffle(X, Y, random_state=seed) 
+    return X, Y, df_response
+
+def tokenizer_save_and_load(tokenizer=None):
+    if tokenizer:
+        file_ = open(tokenizer_weights,'wb')
+        pickle.dump(tokenizer, file_, protocol=pickle.HIGHEST_PROTOCOL)
+        file_.close()
+    else:
+        assert os.path.exists(tokenizer_weights), "Tokenizer Weights doesn't exists. Please Save before load."
+        file_ = open(tokenizer_weights,'rb')
+        tokenizer = pickle.load(file_)
+        file_.close()
+        return tokenizer
+
+def prediction_data(description):
+    tokenizer = tokenizer_save_and_load()
+    processed_description = preprocess_one(description)
+    x_seq = tokenizer.texts_to_sequences([processed_description])
+    x_pad = pad_sequences(x_seq, maxlen=max_length, padding=padding, truncating=trunc_type)
+    return x_pad
+
+def load_prices(df):
+    prices = df[['price{}'.format(i) for i in range(1,n_book_stores + 1)]]
+    prices = prices.apply(lambda x: x.str.strip())
+    prices = prices.apply(lambda x: x.str.replace('$', ''))
+    prices = prices.apply(lambda x: x.str.replace(',', '').astype(float)).values
+    stores = df[websites].values
+
+def load_category_df(df_response, pred_category):
+    df_response['Category'] = df_response['Category'].str.strip()
+    df_response['Category'] = df_response['Category'].str.lower()
+    df_response = df_response[df_response['Category'] == pred_category]
+    return df_response
